+++ NetFabric.Hyperlinq/Conversion/ToList/ToList.Enumerable.cs
@@ -36,7 +36,7 @@ namespace NetFabric.Hyperlinq
             return list;
         }
 
-        static List<TSource> ToList<TEnumerable, TEnumerator, TSource>(this TEnumerable source, Func<TSource, long, bool> predicate)
+        static List<TSource> ToList<TEnumerable, TEnumerator, TSource>(this TEnumerable source, Func<TSource, int, bool> predicate)
             where TEnumerable : IEnumerable<TSource>
             where TEnumerator : IEnumerator<TSource>
         {
