+++ NetFabric.Hyperlinq/Element/First/First.Array.cs
@@ -26,7 +26,7 @@ namespace NetFabric.Hyperlinq
             return ref source[0];
         }
 
-        public static ref TSource First<TSource>(this TSource[] source, Func<TSource, long, bool> predicate)
+        public static ref TSource First<TSource>(this TSource[] source, Func<TSource, int, bool> predicate)
         {
             if (predicate is null) ThrowHelper.ThrowArgumentNullException(nameof(predicate));
 
@@ -40,7 +40,7 @@ namespace NetFabric.Hyperlinq
             return ref source[0];
         }
 
-        public static ref TSource First<TSource>(this TSource[] source, Func<TSource, long, bool> predicate, out int index) 
+        public static ref TSource First<TSource>(this TSource[] source, Func<TSource, int, bool> predicate, out int index) 
         {
             if (predicate is null) ThrowHelper.ThrowArgumentNullException(nameof(predicate));
 
@@ -75,7 +75,7 @@ namespace NetFabric.Hyperlinq
             return ref Default<TSource>.Value;
         }
 
-        public static ref readonly TSource FirstOrDefault<TSource>(this TSource[] source, Func<TSource, long, bool> predicate)
+        public static ref readonly TSource FirstOrDefault<TSource>(this TSource[] source, Func<TSource, int, bool> predicate)
         {
             if (predicate is null) ThrowHelper.ThrowArgumentNullException(nameof(predicate));
 
@@ -88,7 +88,7 @@ namespace NetFabric.Hyperlinq
             return ref Default<TSource>.Value;
         }
 
-        public static ref readonly TSource FirstOrDefault<TSource>(this TSource[] source, Func<TSource, long, bool> predicate, out int index) 
+        public static ref readonly TSource FirstOrDefault<TSource>(this TSource[] source, Func<TSource, int, bool> predicate, out int index) 
         {
             if (predicate is null) ThrowHelper.ThrowArgumentNullException(nameof(predicate));
