+++ NetFabric.Hyperlinq/Element/First/First.Enumerable.cs
@@ -59,7 +59,7 @@ namespace NetFabric.Hyperlinq
             return (ElementResult.Empty, default);
         }
 
-        public static (long Index, TSource Value) TryFirst<TEnumerable, TEnumerator, TSource>(this TEnumerable source, Func<TSource, long, bool> predicate) 
+        public static (int Index, TSource Value) TryFirst<TEnumerable, TEnumerator, TSource>(this TEnumerable source, Func<TSource, int, bool> predicate) 
             where TEnumerable : IEnumerable<TSource>
             where TEnumerator : IEnumerator<TSource>
         {
@@ -67,7 +67,7 @@ namespace NetFabric.Hyperlinq
             {
                 checked
                 {           
-                    for (var index = 0L; enumerator.MoveNext(); index++)
+                    for (var index = 0; enumerator.MoveNext(); index++)
                     {
                         if (predicate(enumerator.Current, index))
                             return (index, enumerator.Current);
@@ -75,7 +75,7 @@ namespace NetFabric.Hyperlinq
                 }
             }        
 
-            return ((long)ElementResult.Empty,  default);
+            return ((int)ElementResult.Empty,  default);
         }
     }
 }
