+++ NetFabric.Hyperlinq/Quantifier/All/All.Array.cs
@@ -17,7 +17,7 @@ namespace NetFabric.Hyperlinq
             return index == length;
         }
 
-        public static bool All<TSource>(this TSource[] source, Func<TSource, long, bool> predicate)
+        public static bool All<TSource>(this TSource[] source, Func<TSource, int, bool> predicate)
         {
             if (predicate is null) ThrowHelper.ThrowArgumentNullException(nameof(predicate));
