+++ NetFabric.Hyperlinq/Quantifier/All/All.Enumerable.cs
@@ -22,7 +22,7 @@ namespace NetFabric.Hyperlinq
             return true;
         }
         
-        public static bool All<TEnumerable, TEnumerator, TSource>(this TEnumerable source, Func<TSource, long, bool> predicate)
+        public static bool All<TEnumerable, TEnumerator, TSource>(this TEnumerable source, Func<TSource, int, bool> predicate)
             where TEnumerable : IEnumerable<TSource>
             where TEnumerator : IEnumerator<TSource>
         {
@@ -30,7 +30,7 @@ namespace NetFabric.Hyperlinq
 
             using (var enumerator = (TEnumerator)source.GetEnumerator())
             {
-                var index = 0L;
+                var index = 0;
                 checked
                 {
                     while (enumerator.MoveNext())
