+++ NetFabric.Hyperlinq/Quantifier/All/All.ReadOnlyCollection.cs
@@ -24,7 +24,7 @@ namespace NetFabric.Hyperlinq
             return true;
         }
 
-        public static bool All<TEnumerable, TEnumerator, TSource>(this TEnumerable source, Func<TSource, long, bool> predicate)
+        public static bool All<TEnumerable, TEnumerator, TSource>(this TEnumerable source, Func<TSource, int, bool> predicate)
             where TEnumerable : IReadOnlyCollection<TSource>
             where TEnumerator : IEnumerator<TSource>
         {
